+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Changelog for Cldr_Messages v0.11.0
+
+This is the changelog for Cldr_Messages v0.11.0 released on ______, 2021.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-cldr/cldr_messages/tags)
+
+### Enhancements
+
+* Add `Cldr.Message.binding/1` to extract the names of variable bindings for a message
+
 # Changelog for Cldr_Messages v0.10.0
 
 This is the changelog for Cldr_Messages v0.10.0 released on April 8th, 2021.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-cldr/cldr_messages/tags)